added new endpoints, fixed test script GET /api/length/{name} POST /api/change_write_key/{name} POST /api/change_read_key/{name}

+++ dbhelper.py
@@ -1,16 +1,17 @@
 import sqlite3 as sq3
-db = sq3.connect("queue.db")
-c = db.cursor()
-c.execute("drop table queue_manager")
-c.execute('''
-    create table if not exists queue_manager (
-          ID integer primary key not null,
-          len integer not null,
-          name text not null,
-          delete_secret text,
-          write_secret text,
-          read_secret text,
-          timestamp integer)
-''')
-db.commit()
-c.close()
+if "y" in input("this will drop queue_manager\n"):
+    db = sq3.connect("queue.db")
+    c = db.cursor()
+    c.execute("drop table queue_manager")
+    c.execute('''
+        create table if not exists queue_manager (
+            ID integer primary key not null,
+            len integer not null,
+            name text not null,
+            delete_secret text,
+            write_secret text,
+            read_secret text,
+            timestamp integer)
+    ''')
+    db.commit()
+    c.close()
